+++ config/services.yaml
@@ -19,12 +19,12 @@ parameters:
     tricks_img_directory: '%kernel.project_dir%/public/uploads/images/tricks'
     avatars_directory: '%kernel.project_dir%/public/uploads/images/users'
     # Number ot tricks displayed on homepage load.
-    homepage_tricks_initial_number : 5
+    homepage_tricks_initial_number: 5
     # Number ot tricks loaded when clicking the "load mode" button.
-    homepage_tricks_load_limit : 5
+    homepage_tricks_load_limit: 5
     # Number ot messages loaded on the trick details page
     # when clicking the "load mode" button.
-    trick_page_messages_load_limit : 10
+    trick_page_messages_load_limit: 10
 
 services:
     # default configuration for services in *this* file
@@ -34,10 +34,11 @@ services:
         public: false       # Allows optimizing the container by removing unused services; this also means
                             # fetching services directly from the container via $container->get() won't work.
                             # The best practice is to be explicit about your dependencies anyway.
-        bind:
-            $homeTricksInitNbr: '%homepage_tricks_initial_number%'
-            $homeTricksLoadLimit: '%homepage_tricks_load_limit%'
-            $trickPageMsgLimit: '%trick_page_messages_load_limit%'
+        #bind:
+        #    $homeTricksInitNbr: '%homepage_tricks_initial_number%'
+        #    $homeTricksLoadLimit: '%homepage_tricks_load_limit%'
+        #    $trickPageMsgLimit: '%trick_page_messages_load_limit%'
+        
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
     App\:
@@ -52,6 +53,14 @@ services:
 
     # add more service definitions when explicit configuration is needed
     # please note that last definitions always *replace* previous ones
+
+    App\Controller\TrickController:
+        #resource: '../src/Controller/TrickController.php'
+        arguments:
+            $homeTricksInitNbr: '%homepage_tricks_initial_number%'
+            $homeTricksLoadLimit: '%homepage_tricks_load_limit%'
+            $trickPageMsgLimit: '%trick_page_messages_load_limit%'
+
 
     #App\Entity\Media:
     #    calls:
