+++ src/DataFixtures/Data/ImagesData.php
@@ -1,0 +1,237 @@
+<?php
+// src/DataFixtures/Data/imagesData.php
+
+namespace App\DataFixtures\Data;
+
+class ImagesData
+{
+    public function getData()
+    {
+        $imagesDetailsArray = [
+            'mute' => [
+                [
+                    'file_name' => 'mute-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'How to do a Mute Grab',
+                    'alt' => 'How to do a Mute Grab',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'mute-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Mute Grab',
+                    'alt' => 'Mute Grab',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+                [
+                    'file_name' => 'mute-3-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Mute Grab close-up',
+                    'alt' => 'Mute Grab close-up',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+                [
+                    'file_name' => 'mute-4-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Mute Grab',
+                    'alt' => 'Mute Grab',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+            'sad' => [
+                [
+                    'file_name' => 'sad-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'How to do a Mute Grab',
+                    'alt' => 'How to do a Mute Grab',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'sad-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Mute Grab',
+                    'alt' => 'Mute Grab',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+            'indy' => [
+                [
+                    'file_name' => 'indy-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Indy Grab',
+                    'alt' => 'Indy Grab',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+                [
+                    'file_name' => 'indy-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Indy close-up',
+                    'alt' => 'Indy close-up',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'indy-3-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Indy upside-down',
+                    'alt' => 'Indy upside-down',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+            'japan-air' => [
+                [
+                    'file_name' => 'japan-air-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Japan Air Grab',
+                    'alt' => 'Japan Air Grab',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'japan-air-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Japan Air',
+                    'alt' => 'Japan Air',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+            'backside-triple-cork-1440' => [
+                [
+                    'file_name' => 'backside-triple-cork-1440-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Backside Triple Cork 1440',
+                    'alt' => 'Backside Triple Cork 1440',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+            ],
+            'front-flip' => [
+                [
+                    'file_name' => 'front-flip-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Front flip detailed',
+                    'alt' => 'Front flip detailed',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+                [
+                    'file_name' => 'front-flip-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Front flip',
+                    'alt' => 'Front flip',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'front-flip-3-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Front flip detailed',
+                    'alt' => 'Front flip detailed',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+            'backflip' => [
+                [
+                    'file_name' => 'backflip-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Backflip',
+                    'alt' => 'Backflip',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'backflip-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Backflip detailed',
+                    'alt' => 'Backflip detailed',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+            'nose-slide' => [
+                [
+                    'file_name' => 'nose-slide-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Nose slide',
+                    'alt' => 'Nose slide',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'nose-slide-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Nose slide',
+                    'alt' => 'Nose slide',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+            'tail-slide' => [
+                [
+                    'file_name' => 'tail-slide-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Tail slide',
+                    'alt' => 'Tail slide',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+                [
+                    'file_name' => 'tail-slide-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Tail slide',
+                    'alt' => 'Tail slide',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+            ],
+            'method-air' => [
+                [
+                    'file_name' => 'method-air-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Method Air',
+                    'alt' => 'Method Air',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'method-air-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Method Air',
+                    'alt' => 'Method Air',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+            'stalefish' => [
+                [
+                    'file_name' => 'stalefish-1-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Stalefish',
+                    'alt' => 'Stalefish',
+                    'file_type' => 0,
+                    'default_cover' => 1,
+                ],
+                [
+                    'file_name' => 'stalefish-2-1920x1080',
+                    'file_url' => 'jpg',
+                    'title' => 'Stalefish',
+                    'alt' => 'Stalefish',
+                    'file_type' => 0,
+                    'default_cover' => 0,
+                ],
+            ],
+        ];
+
+        return $imagesDetailsArray;
+    }
+}
