+++ src/DataFixtures/Data/VideosData.php
@@ -1,0 +1,187 @@
+<?php
+// src/DataFixtures/Data/VideosData.php
+
+namespace App\DataFixtures\Data;
+
+class VideosData
+{
+    public function getData()
+    {
+        $videosDetailsArray = [
+            'mute' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/k6aOWf0LDcQ" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'How to do a Mute Grab',
+                    'alt' => 'How to do a Mute Grab video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/J-ODWe9wm0g" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => '180 Mute Grab',
+                    'alt' => '180 Mute Grab video',
+                    'file_type' => 1,
+                ],
+            ],
+            'sad' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/KEdFwJ4SWq4" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'How to Grab Sad Air',
+                    'alt' => 'How to Grab Sad Air video',
+                    'file_type' => 1,
+                ],
+            ],
+            'indy' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/iKkhKekZNQ8" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'How to Indy Grab',
+                    'alt' => 'How to Indy Grab video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/6QsLhWzXGu0" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Indy Grab',
+                    'alt' => 'Indy Grab video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/-iLTNuWKmeY" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Snowboard Trick Indy',
+                    'alt' => 'Snowboard Trick Indy video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/bkUDlpMfiv4" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Indy Grab Trick Snowboarding Slowmotion',
+                    'alt' => 'Indy Grab Trick Snowboarding Slowmotion video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/aiDtbEJNamQ" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Snowboard straight air indy',
+                    'alt' => 'Snowboard straight air indy video',
+                    'file_type' => 1,
+                ],
+            ],
+            'japan-air' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/jH76540wSqU" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'How to Grab Japan',
+                    'alt' => 'How to Grab Japan video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/I7N45iRPrhw" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'How To Japan',
+                    'alt' => 'How To Japan video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/CzDjM7h_Fwo" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'How To Japan Grab',
+                    'alt' => 'How To Japan Grab video',
+                    'file_type' => 1,
+                ],
+            ],
+            'backside-triple-cork-1440' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/cWWk4Vo3buU" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Backside Triple Cork 1440',
+                    'alt' => 'Backside Triple Cork 1440 video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/4ekuyDYBrz4" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Mark McMorris Triple Cork 1440 Deconstructed',
+                    'alt' => 'Mark McMorris Triple Cork 1440 Deconstructed video',
+                    'file_type' => 1,
+                ],
+            ],
+            'front-flip' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/gMfmjr-kuOg" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Front flip',
+                    'alt' => 'Front flip video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/64flAiRA0i0" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Tame Dog (Front Flip)',
+                    'alt' => 'Tame Dog (Front Flip) video',
+                    'file_type' => 1,
+                ],
+            ],
+            'backflip' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/Yz4brafqk5A" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Backflip',
+                    'alt' => 'Backflip video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/W853WVF5AqI" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Le Backflip en détail',
+                    'alt' => 'Le Backflip en détail video',
+                    'file_type' => 1,
+                ],
+            ],
+            'nose-slide' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/oAK9mK7wWvw" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Le Nose slide en détail',
+                    'alt' => 'Le Nose slide en détail video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/7AB0FZWyrGQ" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Nose slide',
+                    'alt' => 'Nose slide video',
+                    'file_type' => 1,
+                ],
+            ],
+            'tail-slide' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/PxYKuE_SZec" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Tailslide 270 on snow',
+                    'alt' => 'Tailslide 270 on snow video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/KqSi94FT7EE" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Nose & Tail Slide',
+                    'alt' => 'Nose & Tail Slide video',
+                    'file_type' => 1,
+                ],
+            ],
+            'method-air' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/_Cfssjuv0Zg" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Method Air',
+                    'alt' => 'Method Air video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe src="http://www.zapiks.fr/index.php?action=playerIframe&media_id=99159&width=640&height=360&autoStart=false&language=fr" style="position : absolute; top : 0; left : 0; width : 100%; height : 100%;" frameborder="0" scrolling="no" allowfullscreen></iframe>',
+                    'title' => 'Method Air au ralenti',
+                    'alt' => 'Method Air au ralenti video',
+                    'file_type' => 1,
+                ],
+            ],
+            'stalefish' => [
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/rDzm-lkFAI4" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Stalefish Grab on Snow',
+                    'alt' => 'Stalefish Grab on Snow video',
+                    'file_type' => 1,
+                ],
+                [
+                    'file_url' => '<iframe width="560" height="315" src="https://www.youtube.com/embed/0Oez89EoE_c" frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>',
+                    'title' => 'Stalefish',
+                    'alt' => 'Stalefish video',
+                    'file_type' => 1,
+                ],
+            ],
+        ];
+
+        return $videosDetailsArray;
+    }
+}
